+++ tests/odbc/test-odbc-mysql.cpp
@@ -7,97 +7,14 @@
 
 #include "soci/soci.h"
 #include "soci/odbc/soci-odbc.h"
-#include "common-tests.h"
+#include "mysql/test-mysql.h"
 #include <iostream>
 #include <string>
 #include <ctime>
 #include <cmath>
 
-using namespace soci;
-using namespace soci::tests;
-
 std::string connectString;
 backend_factory const &backEnd = *soci::factory_odbc();
-
-// DDL Creation objects for common tests
-struct table_creator_one : public table_creator_base
-{
-    table_creator_one(session & sql)
-        : table_creator_base(sql)
-    {
-        sql << "create table soci_test(id integer, val integer, c char, "
-                 "str varchar(20), sh int2, ul numeric(20), d float8, "
-                 "num76 numeric(7,6), "
-                 "tm datetime, i1 integer, i2 integer, i3 integer, "
-                 "name varchar(20))";
-    }
-};
-
-struct table_creator_two : public table_creator_base
-{
-    table_creator_two(session & sql)
-        : table_creator_base(sql)
-    {
-        sql  << "create table soci_test(num_float float8, num_int integer,"
-                     " name varchar(20), sometime datetime, chr char)";
-    }
-};
-
-struct table_creator_three : public table_creator_base
-{
-    table_creator_three(session & sql)
-        : table_creator_base(sql)
-    {
-        sql << "create table soci_test(name varchar(100) not null, "
-            "phone varchar(15))";
-    }
-};
-
-struct table_creator_for_get_affected_rows : table_creator_base
-{
-    table_creator_for_get_affected_rows(session & sql)
-        : table_creator_base(sql)
-    {
-        sql << "create table soci_test(val integer)";
-    }
-};
-
-//
-// Support for SOCI Common Tests
-//
-
-class test_context : public test_context_base
-{
-public:
-    test_context(backend_factory const &backEnd,
-                std::string const &connectString)
-        : test_context_base(backEnd, connectString) {}
-
-    table_creator_base * table_creator_1(session& s) const
-    {
-        return new table_creator_one(s);
-    }
-
-    table_creator_base * table_creator_2(session& s) const
-    {
-        return new table_creator_two(s);
-    }
-
-    table_creator_base * table_creator_3(session& s) const
-    {
-        return new table_creator_three(s);
-    }
-
-    table_creator_base * table_creator_4(session& s) const
-    {
-        return new table_creator_for_get_affected_rows(s);
-    }
-
-    std::string to_date_time(std::string const &datdt_string) const
-    {
-        return "\'" + datdt_string + "\'";
-    }
-};
 
 int main(int argc, char** argv)
 {
